Dumb proxy seems to work
No authentication yet, but the basic HTTPD proxy seems to work just
fine. At the moment httpd-proxy-public is just a ClusterIP so you still
need to port-forward to it. When the forward is in place the connections
are proxied to chp which in turn proxies to the single user pods.

In this configuration I can log in to the dummy authenticator, a volume
(EBS) is created and the single user pod fires up without any problem.

We still need

   1. SSL
   2. Shibboleth
   3. Image customizations
   4. ...

N.B. We have to set the c.JupyterHub.base_url via hub.configExtra at the
moment. This should probably be exposed in Values when we are ready for
production.

Signed-off-by: Ian Allison <iana@pims.math.ca>