SSL added via cert-manager
See the bitnami docs on on setting up [ingress with
tls-letsencrypt](https://docs.bitnami.com/kubernetes/how-to/secure-kubernetes-services-with-ingress-tls-letsencrypt/).
This commit allows us to have a load balanced address with a letsencrypt
(prod) certificate.

I'm not sure of the ordering of the various components, but the current
setup (working) was deployed by

1. Deploy this chart without the ingress.yaml (hub namespace)
2. Do step 1 of the linked tutorial and install stable/nginx-ingress
(hub namespace)
3. Set up a CNAME record pointing to the resulting LoadBalancer resource
DNS record
4. Add the httpd-proxy-ingress resource without the TLS config (hub namespace)
5. install the cert-manager chart (cert-manager namespace)
6. Update the ingress.yaml to include the TLS config (hub namespace)
7. Update the httpd-proxy-ingress resource to include the TLS config

Wait a while for the certificate to be issued and the nginx ingress to
be reloaded. Check that you get a valid cert for external https
connenctions.

Signed-off-by: Ian Allison <iana@pims.math.ca>